requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Werkzeug==3.0.3
lxml==5.2.1
Flask-Mail==0.10.0
Flask-APScheduler==1.13.1
Flask-Login==0.6.3
